move test_load_static_earth_relief to test_helpers.py
