+++ pygmt/tests/test_datasets_samples.py
@@ -3,7 +3,6 @@ Test basic functionality for loading sample datasets.
 """
 import pandas as pd
 import pytest
-import xarray as xr
 from pygmt.datasets import (
     load_fractures_compilation,
     load_hotspots,
@@ -15,7 +14,6 @@ from pygmt.datasets import (
     load_usgs_quakes,
 )
 from pygmt.exceptions import GMTInvalidInput
-from pygmt.helpers.testing import load_static_earth_relief
 
 
 def test_load_sample_invalid():
@@ -147,16 +145,3 @@ def test_hotspots():
         "place_name",
     ]
     assert isinstance(data, pd.DataFrame)
-
-
-def test_load_static_earth_relief():
-    """
-    Check that @static_earth_relief.nc loads without errors.
-    """
-    data = load_static_earth_relief()
-    assert data.dims == ("lat", "lon")
-    assert data.shape == (14, 8)
-    assert data.min() == 190
-    assert data.max() == 981
-    assert data.median() == 467
-    assert isinstance(data, xr.DataArray)
